version = "0.0.4" backbone_dtype: str = "int4"

+++ llm_studio/python_configs/text_causal_language_modeling_config.py
@@ -292,7 +292,7 @@ class ConfigNLPCausalLMArchitecture(DefaultConfig):
     reward_model_class: Any = text_reward_model.RewardModel
     pretrained: bool = True
 
-    backbone_dtype: str = "float16"
+    backbone_dtype: str = "int4"
     gradient_checkpointing: bool = True
     force_embedding_gradients: bool = False
     intermediate_dropout: float = 0
